+++ RELEASES.md
@@ -139,3 +139,13 @@ Based on:
 - [typescript v0.2.4-alpha.9] .
 ### Releases
 - [NPM v0.2.4-alpha.9] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.9 - .
+
+## 2025-03-20 09:24:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.521.0 (2.558.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.4-alpha.10] .
+### Releases
+- [NPM v0.2.4-alpha.10] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.10 - .